NN-708 #time 2h feature: Integrate API  - Populated API response to PaymentSuccessfulActivity.  - Fixed issues with UI.  - Tested the changes made.

+++ app/src/main/java/com/afrimax/paymaart/data/model/PersonTransactions.kt
@@ -1,30 +1,20 @@
 package com.afrimax.paymaart.data.model
+
 import com.google.gson.annotations.SerializedName
-import java.text.SimpleDateFormat
-import java.util.Date
-import java.util.Locale
 
 
 data class PersonTransactions(
-    @SerializedName("data")
-    val transactions: List<Transaction>,
-    @SerializedName("success_status")
-    val successStatus: Boolean,
-    @SerializedName("total_count")
-    val totalCount: Int
+    @SerializedName("data") val transactions: List<Transaction>,
+    @SerializedName("success_status") val successStatus: Boolean,
+    @SerializedName("total_count") val totalCount: Int
 )
 
 data class Transaction(
-    @SerializedName("created_at")
-    val createdAt: Long,
-    @SerializedName("note")
-    val note: String? = null,
-    @SerializedName("total_amount")
-    val totalAmount: String,
-    @SerializedName("transaction_id")
-    val transactionId: String,
-    @SerializedName("type")
-    val transactionType: String,
+    @SerializedName("created_at") val createdAt: Double,
+    @SerializedName("note") val note: String? = null,
+    @SerializedName("total_amount") val totalAmount: String,
+    @SerializedName("transaction_id") val transactionId: String,
+    @SerializedName("type") val transactionType: String,
     val showDate: Boolean = false,
 )
